refactor(blinc_cn): extract shared element wrapper for form and field
